requirements.txt
alembic 
attrs 
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.0.3
contourpy 
cycler
docopt==0.6.2
Flask 
Flask-Admin 
Flask-Login 
Flask-Migrate 
Flask-SQLAlchemy
Flask-WTF 
fonttools==4.25.0
greenlet 
gunicorn==20.1.0
idna==3.4
importlib-metadata 
importlib-resources
iniconfig
itsdangerous 
Jinja2 
kiwisolver 
Mako 
MarkupSafe 
matplotlib 
munkres==1.1.4
numpy
packaging
Pillow==9.4.0
pipreqs==0.4.13
pluggy
ply==3.11
py
pyparsing
PyQt5-sip==12.11.0
pytest==7.1.2
python-dateutil
requests==2.31.0
sip
six
SQLAlchemy 
toml
tomli
tornado 
typing_extensions
urllib3==2.0.3
Werkzeug==2.3 
WTForms
yarg==0.1.9
zipp 
